does count for words, fix some time complexity issues
to do: blank square, possibly make faster? 
some problems with returning words that aren't words, attributed to them being defined as symbols in the dictionary. user has to remove from enchant dictionary to fix, the only way to fix it in my program alone would be to generate an executable. will do when finished.